Fix scrollBodyTo won't work with floating elements
__Method__
Create a HTML document with a floating element that extends beyond the bottom of the page.
Call scrollBodyTo('bottom').

__Expected Behavior__
The window should scroll to the bottom of the page.

__Actual Behavior__
The window does not scroll at all.

__Cause__
The floating element breaks the javascript scrollIntoView function.

__Resolution__
I changed the method to use window.scrollTo instead, which works file with floating elements.