Moved phpunit to dev dependencies (#77)
__Problem__
PHPUnit was declared as a regular dependency, which meant that any
client consuming this composer package would end up getting phpunit
installed. More importantly, the version requirements for phpunit
imposed by this package would have to be adhered to by clients that
consumed this package.

This causes particular issues with projects that require stable
dependencies, because the phpunit version required by this package
is not stable.

__Fix__
I moved the phpunit dependency to the dev dependencies.