Update to new BDD standards (#118)
* Bump PHP version to 5.6.31

* Convert Spinner trait to class

* Move TypeCaster trait to Medology\Behat namespace

The TypeCaster trait is useful outside of Mink, so we should not keep it within the FlexibleMink namespace, as this implies that it's a trait specificly for Mink.

* Convert ScreenShotContext trait to class

* Fix bin/php does not pass args properly

Problem:
Arguments are not passed to the PHP container propertly. For example, the following command:

`behat --format-settings='{"expand": true}'`

Results in the space between `:` and `true}` separating them into different args, causing Behat to interpret the `true}` as the --paths argument.

Fix:
I wrapped the `$@` of the php script in double quotes so that the arguments structure will be preserved.

* Fix psr-0 in composer.json

* Convert AlertContext trait to class

* Convert JavaScriptContext trait to class

* Delete MinkContextInterface

* Convert StoreContext trait to class

Note: This commit add gatherContexts() method to many traits. Since the use case for these is to mix more than one into a single Behat context, this is going to break the code.

However, this is a temporary situation, as all of these traits are going to be converted to classes in subsequent commits.

* Convert CsvContext trait to class

* Convert TableContext trait to class

* Removed Given phrases from assertions in TableContext

* Convert WebDownloadContext trait to class

* Deleted ContainerContext

This is better served by the QualityAssurance context (not part of this project).

* Move FlexibleMink to Medology\Behat\Mink namespace

* Moved ParalellWorker to Medology namespace