Reduce duplication via UsesStoreContext trait
Problem:
Each context that wishes to access the StoreContext instance needs to implement gathersContexts. This causes duplication in the codebase, and even more so in projects that leverage Flexible Mink.

Solution:
I added a trait called UsesStoreContext that gathers the StoreContext instance and makes it available via the $this->storeContext property.