Reduce duplication via UsesFlexibleContext trait
Problem:
Each context that wishes to access the FlexibleContext instance needs to implement gathersContexts. This causes duplication in the codebase, and even more so in projects that leverage Flexible Mink.

Solution:
I added a trait called UsesFlexibleContext that gathers the FlexibleContext instance and makes it available via the $this->flexibleContext property.