Update Composer bin to 1.5.2 (#124)
Problem:
The composer script is actually an alias for a docker image that contains composer. Although this works, it is not necessary as we already have a PHP image in the project.

Solution:
In addition to updated the composer version, I modified the project so that the composer binary is not in version control. Instead it is installed via a command in init_project. The composer file that is installed is the normal phar, and is simply executed against the bin/php Docker image since the `#!/usr/bin/env php` at the beginning of the composer file will resolve to `./bin/php` if `./bin` is sufficiently high in the users PATH list (which it should be).