Remove retry from "Then" steps. (#127)
* Replace waitFor calls with BehatRetryExtension

* Remove BehatRetryExtension and tests

The team decided that consumers of the FlexibleMink package should be free to choose whether or not they want the retry functionality. So it has been removed from this package and can be added by the consumer if necessary.

* StyleCI Fixes

* Fix circle.yml assume merge target is master

Problem:
The circle.yml config file is attempting to rebase onto origin/master. However PR's are not necessarily against this branch. When they are not, then the merge will either cause a conflict, or will create an invalid codebase against which it runs the tests.

Fix:
I could not find a way to determine what the merge target was from within the CircleCI platorm. Their documentation for the available environmental variables does not mention anything like this: https://circleci.com/docs/1.0/environment-variables/

For now, I have simply disabled the rebasing. I will reach out to our CircleCI rep and see if there's a way to determine the merge target.

* Update docs to reflect no auto-spin

* Remove assertPageContainsTextTemporarily() method

This method is highly susceptible to race conditions, and there are more reliable ways to assert the expected system behavior. As such, I have removed the method from the package.