Fix Carbon instances are not formatted properly by injectStoredValues (#155)
Problem:
Eloquent uses a static dateFormat property on the class which will cause the Carbon instances to be formatted when the model is converted to an array or JSON. If the Carbon instance is converted to a string via PHP, the dateFormat property is not going to be used. This can cause problems because comparing a string Carbon instance locally to one received from the server will result in different formatting.

Fix:
I implemented a method that will determine if the property being read is a DateTime object, and if so, check for the existence of a dateFormat property on the host object. If one exists, the DateTime instance will be formatted as per the dateFormat property.