Cherry pick missing commits from 1.0 branch (#168)
* Fixing issue with nested table check (#125)

* Creating a test that ensures that the `TableContext::assertTableWithStructureExists()` method works with
nested tables.

* The `TableContext::assertTableWithStructureExists()` method was not working whenever a table was nested in
another because when retrieving rows or columns, it would attempt to get all
columns within the parent table which would then include rows and columns of children tables.

This commit will ensure that whenever tables are checked, rows within one level of depth are returned whenever
retrieving columns or rows in this method.

* Adding injectStoredValues to Method (#126)

* Adding injectStoredValues to the usage of method `TableContext::assertTableWithStructureExists()`.

* Since this is part of another fix as a whole, including this item that was missed in my previous pull request.

* Add Check if Field is visible (#128)

* Addition of step definition to allow check if a field is visible or not when is it supposed to be visible or not.

* Apply fixes from StyleCI

* PR Review Changes requested by Johnnie

* Apply fixes from StyleCI

* Field Visibility Tests - Behat Fix (#129)

* Addition of test to proof that assert field visibility works.
fix on the behat issue not being able to test a specific step or feature using the command `behat --name=""`

* Donald PR review change.

* Ramsey PR Changes.

* Specify format when injecting values (#156)

* Specify format when injecting values

Problem:
The injectStoredValues method is simplistic in that it injects the stored value as is without any context. For the most part this is fine, but sometimes we have a model that has a DateTime which is being displayed by the presentation tier using a different format. For example, the created_at property of a model might include hours, mins and seconds, but the front-end may decide that they only wish to display year, month and day (probably in standard US format). If the stringified date from the store does not match how the front-end is presenting it, then it's not possible to say something like `I should see (the created at date of the order)`.

Fix:
I implemented the ability to specify a format for the injected value. For example `I should see (the created date of the order as a US date)`. I added a few useful formats, and developers can easily add more as needed.

* Code review changes

* Fix stdClass with dateFormat property causes ReflectionException

Problem:
Using a stdClass as the host object causes a ReflectionException because stdClass is not a real class.

Fix:
Since stdClass cannot have private or protected properties, I used a simple isset to check if the property exists before attempting to use reflection. If the object has a public property, its value will be returned. If it does not, reflection will be used. Either the reflection will succeed and the private or protected property will be returned, or it will fail because the object either has no class, or has no property with the specified name.

* Fix DateTime cannot be cast to string

Problem:
The methods are coded against the DateTime class, even though my need is for them to work with Carbon. When you do actually use them with a real DateTime class, the methods fail when they try to cast the DateTime to a string.

Cause:
The DateTime class has no __toString() method.

Fix:
I modified the code to check if the DateTime object has a __toString() method. If it does (such as in the case of Carbon), it will use it. If it does not, it will fall back to calling format() with the DateTime::ISO8601 standard.

* Added tests for injectStoredValues with format

* Apply StyleCI patch

* Dan's code review changes

* Fix "0" is not converted to int (#157)

Problem:
Providing 0 for a parameter value in a Behat step does not get converted to int.

Cause:
The pattern was recently updated to exclude numbers prefixed with 0 so that zip codes starting with zero would not be converted to int. The change did not take into account the scenario where the only digit is 0.

Fix:
I modified the pattern to convert a single 0 to an integer in addition to all numbers that are two or more digits that do not start with a zero.

* Implement horizontal scrolling (#158)

Requirement:
I need to test a page where the content causes a horizontal scroll bar to be present. While filling out the form the page is automatically scrolled to the right. I need to be able to scroll back to the left to interact with an element that is now outside of the viewport.

Implementation:
I modified the scrollWindowToBody function to allow any valid combination of top, bottom, left right so that we can scroll in any direction as necessary.