Exception cleanup (#170)
* Spinner::waitFor throws specific exception

Problem:
Spinner::waitFor was defaulting to an exception type of "Exception". This makes it very difficult for a caller to catch and handle the problem because it's very difficult to distinguish between this and any other "Exception" (which is EVERY exception).

Fix:
Modified the class to return a new SpinnerTimoutException instead. It still "is-a" Exception, so backwards compatability is maintained. Consumers of the package can tighten up their exception handling at their convenience.

* StoreContext::injectStoredValues throws specific exceptions

Problem:
StoreContext::injectStoredValues was throw various exceptions of type of "Exception". This makes it very difficult for a caller to catch and handle the problem because it's very difficult to distinguish between this and any other "Exception" (which is EVERY exception).

Fix:
Modified the class to return more specific/appropriate exception types instead. They still are Exceptions, so backwards compatability is maintained. Consumers of the package can tighten up their exception handling at their convenience.

* Fix missing throws annotations