Allow registration of Store initialization lambdas (#227)
Problem:
When the StoreContext was a trait in FlexibleMink 1.0, it used a method_exists() check to determine if there was a onStoreInitialized() method, and called it if so. This was used by consumers to register certain nouns that should be available to all tests in the suite, such as "Browser" which would provide the IP Address etc of the client's browser. Since the StoreContext is no longer a trait, this method does not work.

Fix:
I replaced the method_exists() check with a registerInitialization() method that the consumer can pass a lamda to. The lambda will be called immediately after the store is reset between each test scenario, allowing the store to be seeded or modified as necessary.