Fix getListOfAllNodeElementParents includes stopAt node (#341)
* Rename getListOfAllNodeElementParents to getAncestors

* Rename nodeElement to node

* Add failing test

* Add missing param annotation for node param

* Don't modify string case on each while iteration

* Don't call getParent() twice per iteration

* Don't include stopAt node in ancestor list

* Rename nodeElements to nodes

* Make stopAt optional

* Apply StyleCI fixes

* Fix elements with body as parent are never visible

Problem:
When an element has the body as it's parent, nodeIsVisibleInViewport always returns false.

Cause:
The method is telling getAncestors() to stop at body, so the parents list is empty. Then the method is assuming that if the parents list is empty, the element is detached from the DOM.

Fix:
Check the elements parent directly instead of the filtered parents list returned from getAncestors().