Fix stale data returned from getTableFromName (#348)
Problem:
In FlexibleMink 1.0, the getTableFromName() method would return a cached version of the table structure if it was available and the forceFresh parameter was not true. This worked fine for 1.0, but since 2.0 is intended to be used with the automatic spinning of the BehatRetryExtension, it was preventing the retry from ever passing. This was due to the first call returning the fresh table data that did not pass the assertion, and subsequent calls would receive the stale data even if the real table on the page had updated and should pass the assertion.

Fix:
I removed the caching entirely. It is not appropriate for a generic table checking method, especially in an async environment such as Selenium. If the users of FlexibleMink want caching (say, because they are using the sync MinkDriver instead of the async SeleniumDriver), then they can implement caching around FlexibleMink.